move cloud-sync to main
