+++ packages/insomnia/src/main/cloud-sync/core/__tests__/util.test.ts
@@ -1,9 +1,9 @@
 import { createBuilder } from '@develohpanda/fluent-builder';
 import { beforeEach, describe, expect, it } from 'vitest';
 
-import { baseModelSchema, workspaceModelSchema } from '../../../models/__schemas__/model-schemas';
-import { branchSchema, mergeConflictSchema, statusCandidateSchema } from '../../__schemas__/type-schemas';
-import type { StageEntry } from '../../types';
+import { baseModelSchema, workspaceModelSchema } from '../../../../models/__schemas__/model-schemas';
+import { branchSchema, mergeConflictSchema, statusCandidateSchema } from '../../../../sync/__schemas__/type-schemas';
+import type { StageEntry } from '../../../../sync/types';
 import {
   combinedMapKeys,
   compareBranches,
