+++ packages/insomnia/src/main/cloud-sync/core/__tests__/vcs.test.ts
@@ -1,11 +1,11 @@
 import { createBuilder } from '@develohpanda/fluent-builder';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { baseModelSchema, workspaceModelSchema } from '../../../models/__schemas__/model-schemas';
-import { projectSchema } from '../../__schemas__/type-schemas';
-import { shouldIgnoreKey } from '../../ignore-keys';
-import { deterministicStringify } from '../../lib/deterministic-stringify';
-import type { BackendProject } from '../../types';
+import { baseModelSchema, workspaceModelSchema } from '../../../../models/__schemas__/model-schemas';
+import { projectSchema } from '../../../../sync/__schemas__/type-schemas';
+import { shouldIgnoreKey } from '../../../../sync/ignore-keys';
+import { deterministicStringify } from '../../../../sync/lib/deterministic-stringify';
+import type { BackendProject } from '../../../../sync/types';
 import MemoryDriver from '../store/drivers/memory-driver';
 import { chunkArray, VCS } from '../vcs';
