+++ packages/insomnia/src/main/cloud-sync/core/store/hooks/__tests__/compress.test.ts
