+++ packages/insomnia/src/main/cloud-sync/core/store/hooks/compress.ts
