+++ packages/insomnia/src/main/cloud-sync/create-vcs.ts
@@ -1,6 +1,6 @@
 import type { MergeConflict } from '../../sync/types';
-import FileSystemDriver from '../../sync/vcs/store/drivers/file-system-driver';
-import { VCS } from '../../sync/vcs/vcs';
+import FileSystemDriver from './core/store/drivers/file-system-driver';
+import { VCS } from './core/vcs';
 
 export type ConflictHandler = (
   conflicts: MergeConflict[],
