+++ packages/insomnia/src/main/cloud-sync/vcs.ts
@@ -5,11 +5,11 @@ import { app, type WebContents } from 'electron';
 
 import type { RemoteProject } from '~/insomnia-data';
 import { services } from '~/insomnia-data';
+import type { VCS } from '~/main/cloud-sync/core/vcs';
 import { createVCS } from '~/main/cloud-sync/create-vcs';
 import { pullBackendProject } from '~/main/cloud-sync/pull-backend-project';
 import type { BackendProjectWithTeam, MergeConflict } from '~/sync/types';
 import { UserAbortResolveMergeConflictError } from '~/sync/vcs/errors';
-import type { VCS } from '~/sync/vcs/vcs';
 import { invariant } from '~/utils/invariant';
 
 interface SyncInvocationContext {
